aaaarghhh day 9

+++ day_09/computing.h
@@ -8,4 +8,4 @@ expansion expand_disk_map(disk_map m);
 
 void arrange_expansion(expansion ex);
 
-long long int compute_checksum(expansion ex);
+ULLONG compute_checksum(expansion ex);
